refactor: do some refactoring
Changes the name of the image mapa from mapa.png to map.png. This change affected the index.html so it changes the names of the image there. Also changes the name of the css file from styles.css to headerAndFooter.css to have an specific file for its styles.